Move pantry into directory.
