+++ README.md
@@ -150,7 +150,7 @@ colormap.
 
 ```python
 import geovista as gv
-from geovista.pantry import ww3_global_tri
+from geovista.assets.pantry import ww3_global_tri
 import geovista.theme
 
 # Load the sample data.
@@ -188,7 +188,7 @@ from an [FVCOM](http://fvcom.smast.umassd.edu/fvcom/) **unstructured** mesh, as 
 
 ```python
 import geovista as gv
-from geovista.pantry import fvcom_tamar
+from geovista.assets.pantry import fvcom_tamar
 import geovista.theme
 
 # Load the sample data.
@@ -239,7 +239,7 @@ base layer.
 
 ```python
 import geovista as gv
-from geovista.pantry import lam_pacific
+from geovista.assets.pantry import lam_pacific
 import geovista.theme
 
 # Load the sample data.
@@ -283,7 +283,7 @@ base layer.
 import cartopy.crs as ccrs
 
 import geovista as gv
-from geovista.pantry import lam_pacific
+from geovista.assets.pantry import lam_pacific
 import geovista.theme
 
 # Load the sample data.
@@ -325,7 +325,7 @@ Now render a [Met Office LFRic](https://www.metoffice.gov.uk/research/approach/m
 
 ```python
 import geovista as gv
-from geovista.pantry import lfric_sst
+from geovista.assets.pantry import lfric_sst
 import geovista.theme
 
 # Load the sample data.
@@ -364,7 +364,7 @@ using Met Office Unified Model (UM) ORCA2 Sea Water Potential Temperature data, 
 
 ```python
 import geovista as gv
-from geovista.pantry import um_orca2
+from geovista.assets.pantry import um_orca2
 import geovista.theme
 
 # Load sample data.
@@ -402,7 +402,7 @@ Now let's render a [NOAA/NCEI Optimum Interpolation SST](https://www.ncei.noaa.g
 
 ```python
 import geovista as gv
-from geovista.pantry import oisst_avhrr_sst
+from geovista.assets.pantry import oisst_avhrr_sst
 import geovista.theme
 
 # Load sample data.
@@ -442,7 +442,7 @@ model uses hexagonal and pentagonal cells, and is a new dynamical core for
 
 ```python
 import geovista as gv
-from geovista.pantry import dynamico
+from geovista.assets.pantry import dynamico
 import geovista.theme
 
 # Load sample data.
