+++ src/geovista/examples/point_cloud/from_points__orca_cloud_eqc.py
@@ -32,7 +32,7 @@ projection.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import um_orca2_gradient
+from geovista.assets.pantry import um_orca2_gradient
 from geovista.samples import ZLEVEL_SCALE_CLOUD
 import geovista.theme
