+++ src/geovista/examples/rectilinear/from_1d__oisst.py
@@ -30,7 +30,7 @@ NASA Blue Marble base layer is rendered along with Natural Earth coastlines.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import oisst_avhrr_sst
+from geovista.assets.pantry import oisst_avhrr_sst
 import geovista.theme
