+++ src/geovista/examples/scalar_data/earthquakes_wink1.py
@@ -48,7 +48,7 @@ from __future__ import annotations
 from warnings import warn
 
 import geovista as gv
-from geovista.pantry import usgs_earthquakes
+from geovista.assets.pantry import usgs_earthquakes
 import geovista.theme
