+++ src/geovista/examples/unstructured/fesom.py
@@ -31,7 +31,7 @@ coastlines.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import fesom
+from geovista.assets.pantry import fesom
 import geovista.theme
