+++ src/geovista/examples/unstructured/fesom_fouc.py
@@ -32,7 +32,7 @@ projection.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import fesom
+from geovista.assets.pantry import fesom
 import geovista.theme
