+++ src/geovista/examples/unstructured/icon_eqc.py
@@ -33,7 +33,7 @@ from __future__ import annotations
 import matplotlib as mpl
 
 import geovista as gv
-from geovista.pantry import icon_soil
+from geovista.assets.pantry import icon_soil
 import geovista.theme
