+++ src/geovista/examples/unstructured/lam_pacific_moll.py
@@ -31,7 +31,7 @@ projection.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import lam_pacific
+from geovista.assets.pantry import lam_pacific
 import geovista.theme
