+++ src/geovista/examples/unstructured/lfric_sst.py
@@ -30,7 +30,7 @@ base layer is also rendered along with Natural Earth coastlines and a graticule.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import lfric_sst
+from geovista.assets.pantry import lfric_sst
 import geovista.theme
