+++ src/geovista/examples/unstructured/smc_sinu.py
@@ -31,7 +31,7 @@ projection.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import ww3_global_smc
+from geovista.assets.pantry import ww3_global_smc
 import geovista.theme
