+++ src/geovista/examples/unstructured/tri.py
@@ -32,7 +32,7 @@ across the mesh faces/cells.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import ww3_global_tri
+from geovista.assets.pantry import ww3_global_tri
 import geovista.theme
