+++ src/geovista/examples/unstructured/tri_hammer.py
@@ -33,7 +33,7 @@ across the mesh faces/cells.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import ww3_global_tri
+from geovista.assets.pantry import ww3_global_tri
 import geovista.theme
