+++ src/geovista/examples/warp/from_unstructured__lfric_orog.py
@@ -32,7 +32,7 @@ coastlines are rendered.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import lfric_orog
+from geovista.assets.pantry import lfric_orog
 import geovista.theme
