+++ src/geovista/examples/warp/from_unstructured__lfric_orog_eqc.py
@@ -34,7 +34,7 @@ Cylindrical (Plate Carrée) conformal cylindrical projection.
 from __future__ import annotations
 
 import geovista as gv
-from geovista.pantry import lfric_orog
+from geovista.assets.pantry import lfric_orog
 import geovista.theme
 from geovista.transform import transform_mesh
